LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/unhandled-promise-rejections/promise-rejection-events-onerror-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: e
CONSOLE MESSAGE: Error: e2

PASS Throwing inside an unhandledrejection handler invokes the error handler.
